main code
Adds main to the project, which houses the code for executing the example program. 

To run this program, chromedriver must be installed in the same folder as main.py. Chromedriver can be downloaded at https://chromedriver.storage.googleapis.com/index.html

Make sure that the downloaded version of Chromedriver matches Chromes current version.